Software/Python/requirements.txt
pyserial==3.5
SimConnect==0.4.26